2025_tinkoff_ctf/web/brethren/sources/backend/pyproject.toml
[tool.poetry]
name = "t14-3"
version = "0.1.0"
description = ""
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
litestar = "^2.15.1"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.38"
pydantic = "^2.10.6"
dishka = "^1.4.2"
alembic = "^1.15.1"
python-dotenv = "^1.0.1"
python-jose = "^3.4.0"
psycopg = "^3.2.5"
uvloop = "^0.21.0"
httptools = "^0.6.4"
passlib = "^1.7.4"
email-validator = "^2.2.0"



[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
